[TASK] Drop the `AbstractFrontendUserRepository`
After dropping `FrontendUserWithCountry` and its corresponding repository,
there is no need to have the abstract repository class anymore.
