+++ Classes/Domain/Repository/FrontendUserRepository.php
@@ -5,10 +5,38 @@ declare(strict_types=1);
 namespace OliverKlee\FeUserExtraFields\Domain\Repository;
 
 use OliverKlee\FeUserExtraFields\Domain\Model\FrontendUser;
+use TYPO3\CMS\Extbase\Persistence\Repository;
 
 /**
- * @extends AbstractFrontendUserRepository<FrontendUser>
+ * @extends Repository<FrontendUser>
  */
-class FrontendUserRepository extends AbstractFrontendUserRepository
+class FrontendUserRepository extends Repository implements DirectPersistInterface
 {
+    use DirectPersistTrait;
+
+    public function findOneByUsername(string $username): ?FrontendUser
+    {
+        if ($username === '') {
+            return null;
+        }
+
+        $query = $this->createQuery();
+        $query->getQuerySettings()->setRespectStoragePage(false);
+        $query->matching($query->equals('username', $username));
+
+        return $query->execute()->getFirst();
+    }
+
+    public function existsWithUsername(string $username): bool
+    {
+        if ($username === '') {
+            return false;
+        }
+
+        $query = $this->createQuery();
+        $query->getQuerySettings()->setRespectStoragePage(false);
+        $query->matching($query->equals('username', $username));
+
+        return $query->execute()->count() > 0;
+    }
 }
